git: write-tree
