--- completers/git_completer/cmd/write_tree_generated.go
@@ -1,18 +1,0 @@
-package cmd
-
-import (
-	"github.com/spf13/cobra"
-)
-
-var write_treeCmd = &cobra.Command{
-	Use:     "write-tree",
-	Short:   "Create a tree object from the current index",
-	Run:     func(cmd *cobra.Command, args []string) {},
-	GroupID: groups[group_low_level_manipulator].ID,
-}
-
-func init() {
-	write_treeCmd.Flags().Bool("missing-ok", false, "allow missing objects")
-	write_treeCmd.Flags().String("prefix", "", "write tree object for a subdirectory <prefix>")
-	rootCmd.AddCommand(write_treeCmd)
-}
